feat: legacy ui is working

+++ backend/src/validations/schemas/bounds.ts
@@ -2,10 +2,10 @@ import { z } from "zod";
 
 const BOUNDS_QUERY_SCHEMA = z
   .object({
-    minLatitude: z.string().transform((arg) => Number.parseInt(arg)),
-    maxLatitude: z.string().transform((arg) => Number.parseInt(arg)),
-    minLongitude: z.string().transform((arg) => Number.parseInt(arg)),
-    maxLongitude: z.string().transform((arg) => Number.parseInt(arg)),
+    minLatitude: z.string().transform((arg) => Number.parseFloat(arg)),
+    maxLatitude: z.string().transform((arg) => Number.parseFloat(arg)),
+    minLongitude: z.string().transform((arg) => Number.parseFloat(arg)),
+    maxLongitude: z.string().transform((arg) => Number.parseFloat(arg)),
   })
   .or(
     z.object({
